Proposal for restructuring the counts with threshold functionality.
Some aspects:
- organise static helper functions in a separate class
- split implementations for cross table and counts
- let one service implement the resource interface, because
  the services do different things and, hence, should not be
  implementing the same interface. (Probably the service without
  threshold should use different names for the implementing functions).

Note:
To keep the unit tests for the threshold layer working, there
still needs to be separate beans, but ideally I think there
should only be one for counts and one for the cross table, instead
of two for both.
