--- transmart-core-api/src/main/groovy/org/transmartproject/core/config/CountsThresholdResource.groovy
@@ -1,14 +1,0 @@
-package org.transmartproject.core.config
-
-
-interface CountsThresholdResource {
-
-    /**
-     * Get threshold value, below which counts are not available for users
-     * with `COUNTS_WITH_THRESHOLD` access permission.
-     *
-     * @return threshold value
-     */
-    long getPatientCountThreshold()
-
-}
